Add finance module to home-manager
The finance module only contains configuration of ledger-cli for now.
There may be more software to follow in the future, though. This is the
first module to use custom options (in the host-specific) domain to
adjust things like file paths on a host-to-host basis. Ideally every
host then has a host-specific section in its home.nix to take care of
such things. Let's see how this turns out.