Create gui module for home-manager
This contains mostly setup for gui programs like firefox, kitty etc.
Thus it is distinct from the system gui module which installs, e.g., the
desktop environment.

The corresponding old roles are deleted and merged into the new module
to fit the desired new structure. Dotfiles are moved into the dotfiles
directory and might be deleted in the near future as we don't need them
anymore.